+++ projects/form-builder/src/lib/utils/parser.ts
@@ -434,7 +434,9 @@ export class Parser {
       pointer: string
     }
   ) {
-    const pointers = parentArray ? (this.pointers[parentArray.pointer] as any).arrayPointers[parentArray.index] : this.pointers;
+    const pointers = parentArray ?
+      (this.pointers[parentArray.pointer] as any).arrayPointers[parentArray.index] :
+      this.pointers;
     const target = pointers[pointer];
     const control = pointers[pointer].control as FormArray;
 
@@ -473,7 +475,9 @@ export class Parser {
       pointer: string
     }
   ) {
-    const pointers = parentArray ? (this.pointers[parentArray.pointer] as any).arrayPointers[parentArray.index] : this.pointers;
+    const pointers = parentArray ?
+      (this.pointers[parentArray.pointer] as any).arrayPointers[parentArray.index] :
+      this.pointers;
     const target = pointers[pointer];
     const control = pointers[pointer].control as FormArray;
 
@@ -580,22 +584,7 @@ export class Parser {
    * - Handle items or contains as array not object
    */
   private buildArray(base: string, definition: ArrayPropertyDefinition) {
-    if (!definition.items) {
-      return {
-        control: new FormControl([]),
-        ...(definition.items
-          ? {
-              arrayType: definition.items.type,
-              properties: definition.items.properties,
-              required: definition.items.required,
-              validation: {}
-            }
-          : {
-              arrayType: SchemaType.String,
-              validation: {}
-            })
-      };
-    } else {
+    if (definition.items) {
       return {
         arrayType: definition.items.type,
         properties: definition.items.properties,
@@ -603,6 +592,12 @@ export class Parser {
         validation: {},
         control: new FormArray([]),
         arrayPointers: []
+      };
+    } else {
+      return {
+        control: new FormControl([]),
+        arrayType: SchemaType.String,
+        validation: {}
       };
     }
   }