+++ projects/page-builder/src/public-api.ts
@@ -2,5 +2,12 @@
  * Public API Surface of page-builder
  */
 export * from './lib/page-builder.module';
+export * from './lib/inline-editor/inline-editor.module';
 export * from './lib/options.interface';
 export * from './lib/options.token';
+
+/**
+ * Inline editors
+ */
+export * from './lib/inline-editor/directives/image-ie.directive';
+export * from './lib/inline-editor/directives/single-line-ie.directive';