+++ projects/form-ui/CHANGELOG.md
@@ -1,3 +1,57 @@
+## @jaspero/fb-form-ui [1.9.47](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.46...@jaspero/fb-form-ui@1.9.47) (2021-06-05)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.7
+* **@jaspero/fb-tinymce:** upgraded to 1.8.3
+
+## @jaspero/fb-form-ui [1.9.46](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.45...@jaspero/fb-form-ui@1.9.46) (2021-06-04)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.6
+* **@jaspero/fb-tinymce:** upgraded to 1.8.2
+
+## @jaspero/fb-form-ui [1.9.45](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.44...@jaspero/fb-form-ui@1.9.45) (2021-06-04)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.5
+* **@jaspero/fb-tinymce:** upgraded to 1.8.1
+
+## @jaspero/fb-form-ui [1.9.44](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.43...@jaspero/fb-form-ui@1.9.44) (2021-06-03)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/fb-tinymce:** upgraded to 1.8.0
+
+## @jaspero/fb-form-ui [1.9.43](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.42...@jaspero/fb-form-ui@1.9.43) (2021-05-31)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.4
+* **@jaspero/fb-tinymce:** upgraded to 1.7.22
+
 ## @jaspero/fb-form-ui [1.9.42](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-form-ui@1.9.41...@jaspero/fb-form-ui@1.9.42) (2021-05-20)