+++ projects/tinymce/CHANGELOG.md
@@ -1,3 +1,50 @@
+## @jaspero/fb-tinymce [1.8.3](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-tinymce@1.8.2...@jaspero/fb-tinymce@1.8.3) (2021-06-05)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.7
+
+## @jaspero/fb-tinymce [1.8.2](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-tinymce@1.8.1...@jaspero/fb-tinymce@1.8.2) (2021-06-04)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.6
+
+## @jaspero/fb-tinymce [1.8.1](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-tinymce@1.8.0...@jaspero/fb-tinymce@1.8.1) (2021-06-04)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.5
+
+# @jaspero/fb-tinymce [1.8.0](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-tinymce@1.7.22...@jaspero/fb-tinymce@1.8.0) (2021-06-03)
+
+
+### Features
+
+* **tinymce:** added emoticons ([73ab516](https://github.com/Jaspero/schema-forms/commit/73ab51621a7a041ee6e885526091f45f2d7e1c50))
+
+## @jaspero/fb-tinymce [1.7.22](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-tinymce@1.7.21...@jaspero/fb-tinymce@1.7.22) (2021-05-31)
+
+
+
+
+
+### Dependencies
+
+* **@jaspero/form-builder:** upgraded to 3.28.4
+
 ## @jaspero/fb-tinymce [1.7.21](https://github.com/Jaspero/schema-forms/compare/@jaspero/fb-tinymce@1.7.20...@jaspero/fb-tinymce@1.7.21) (2021-05-20)